% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{expect_norm_gq}
\alias{expect_norm_gq}
\title{Calculate the expected value of f(X) with X~N(mu, sigma)}
\usage{
expect_norm_gq(fun, dimensions, center = rep(0, dimensions),
  scale = diag(1, dimensions), settings = defaults.gq(), ...)
}
\arguments{
\item{fun}{Function to integrate}

\item{dimensions}{Number of dimensions of X}

\item{center}{Mean of X}

\item{scale}{Variance of X}

\item{settings}{Settings}

\item{...}{Arguments to f}
}
\value{
Numeric value of the integral
}
\description{
Calculate the expected value of f(X) with X~N(mu, sigma)
}
\examples{
f <- function(x) x^2
expect_norm_gq(f, 1)
}
