% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funsu-package.R
\docType{package}
\name{funsu-package}
\alias{funsu}
\alias{funsu-package}
\title{funsu: Collection of Useful Functions}
\description{
Collection of Useful Functions
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sebastianueckert/funsu}
  \item Report bugs at \url{https://github.com/sebastianueckert/funsu/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Ueckert \email{sebastian.ueckert@gmail.com} (0000-0002-3712-0255)

}
\keyword{internal}
