% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{gq_points_weights}
\alias{gq_points_weights}
\title{Gaussian Quadrature points and weights}
\usage{
gq_points_weights(dimensions = 1, mu = rep(0, dimensions),
  sigma = diag(1, dimensions), settings = settings.gq())
}
\arguments{
\item{dimensions}{Number of dimensions}

\item{mu}{Mean of the normal distribution}

\item{sigma}{Variance covariance matrix of the normal distribution}

\item{settings}{Settings}
}
\value{
List with entries weights and grid_points
}
\description{
The function returns the grid points and weights for the calculation of  means (or variances)
of a function w.r.t. a normal distribution.
}
\examples{
# points and weights for 2 dimensional integration with 3 grid points per dimension
gq_points_weights(2, settings = settings.gq(quad_points = 3))
}
