% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{integrate_gq}
\alias{integrate_gq}
\title{Calculate the expected value of f(X) with X~N(mu, sigma)}
\usage{
integrate_gq(fun, dimensions, center = rep(0, dimensions),
  scale = diag(1, dimensions),
  settings = filter_settings(defaults.agq(), "gq"), ...)
}
\arguments{
\item{fun}{Function to integrate}

\item{dimensions}{Number of dimensions of X}

\item{center}{Mean of X}

\item{scale}{Variance of X}

\item{settings}{Settings}

\item{...}{Arguments to f}
}
\description{
Calculate the expected value of f(X) with X~N(mu, sigma)
}
