% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt.R
\name{irt_expected_cov_matrix}
\alias{irt_expected_cov_matrix}
\title{Calculate expected item covariance matrix for a population}
\usage{
irt_expected_cov_matrix(model, mu = NULL, sigma = NULL,
  settings = settings.irt_expected())
}
\arguments{
\item{model}{A \code{mirt} model}

\item{mu}{The mean of latent variable values}

\item{sigma}{The variance-covariance matrix of the latent variable values}
}
\value{
The variance covariance matrix of the assessment
}
\description{
Calculate expected item covariance matrix for a population
}
\examples{
fit <- mirt::mirt(mirt::Science, 1)
irt_expected_cov_matrix(fit)
}
\seealso{
\code{\link{irt_expected_item_response}}
}
