% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt.R
\name{irt_expected_item_response}
\alias{irt_expected_item_response}
\alias{settings.irt_expected}
\title{Calculate expected item response for a population}
\usage{
irt_expected_item_response(model, mu = NULL, sigma = NULL,
  settings = settings.irt_expected())

settings.irt_expected(gq.quad_points = 5)
}
\arguments{
\item{model}{A \code{mirt} model}

\item{mu}{The mean of latent variable values}

\item{sigma}{The variance-covariance matrix of the latent variable values}
}
\value{
A vector with the expected score for each item
}
\description{
The function calculates the expected item score for a given mirt IRT model and a specified population. If mean and variance are not provided, they are taken from the
mirt model object.
}
\examples{
fit <- mirt::mirt(mirt::Science, 1)
irt_expected_item_response(fit)
}
\seealso{
\code{\link{irt_expected_cov_matrix}}
}
