% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{make_settings}
\alias{make_settings}
\title{Make a settings function}
\usage{
make_settings()
}
\value{
A list
}
\description{
The function simplifies the creation of settings functions which return their arguments as a list. It works by
combining the formals of the calling function with the arguments of the acutal call and returns
the results as a list
}
\examples{
my_settings <- function(opt_1 = 1) make_settings()
settings <- my_settings()
settings2 <- my_settings(opt_1 = 2)
}
