% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{nest_lst_by_name}
\alias{nest_lst_by_name}
\title{Nest list based on element names}
\usage{
nest_lst_by_name(lst)
}
\arguments{
\item{lst}{A list}
}
\value{
A list
}
\description{
The function takes a list and nest all elements that have a name composed with a "."
}
\examples{
nest_lst_by_name(list(gq.opt_1 = 1, gq.opt_2 = 2))
}
