% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{expect_norm_gq}
\alias{expect_norm_gq}
\alias{expect_norm_mc}
\alias{settings.gq}
\alias{settings.mc}
\title{Calculate the expected value of f(X) with X~N(mu, sigma)}
\usage{
expect_norm_gq(fun, dimensions = 1, mu = rep(0, dimensions),
  sigma = diag(1, dimensions), settings = settings.gq(), ...)

expect_norm_mc(fun, dimensions, mu = rep(0, dimensions),
  sigma = diag(1, dimensions), settings = settings.mc(), ...)

settings.gq(quad_points = 5)

settings.mc(n_samples = 1000, seed = 123)
}
\arguments{
\item{fun}{Function to integrate}

\item{dimensions}{Number of dimensions of X}

\item{mu}{Mean of X}

\item{sigma}{Variance of X}

\item{settings}{Settings}

\item{...}{Arguments to f}
}
\value{
Numeric value of the integral
}
\description{
Calculate the expected value of f(X) with X~N(mu, sigma)
}
\section{Functions}{
\itemize{
\item \code{expect_norm_gq}: Calculation using Gaussian quadrature

\item \code{expect_norm_mc}: Calculation using Monte-Carlo sampling
}}

\examples{
f <- function(x) x^2
expect_norm_gq(f)
}
